+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## Unreleased
 
 ### Added
-- Nothing.
+- Added handlebars for handling address template.
 
 ### Changed
 - Removed type of friend.
+- Added international address format when changing address country.
+- Removed clean gulp task from build task.
 
 ### Deprecated
 - Nothing.